Add files via upload
This program reads input values from a file. The values are then processed as data for the amount of
time a system spends idle for a given amount of days. The performance rate, average daily time idle,
and days with maximum and minimum time idle are then written to an output file.